requirements.txt
click>=8.0.0
pyyaml>=6.0
docker>=6.0.0
pytest>=7.0.0